requirements/requirements.txt
Django>=3.0,<4.0
pandas>=1.4.1
django-crispy-forms
crispy_bootstrap5
psycopg2>=2.8
elsapy
pysocks
django-crontab